+++ src/curses/nc_cyclic_buffer.h
@@ -8,11 +8,11 @@
 #include "curses/nc_window.h"
 
 void nc_cyclic_text_write(NcmBuffer *output, char *string,
-                          int32 string_len, int64 *start_pos,
+                          int32 string_len, int32 *start_pos,
                           int32 width, char *separator,
                           int32 separator_len, bool scrolling_enabled);
 void nc_cyclic_buffer_write(NcBuffer *buffer, NcWindow *window,
-                            int64 *start_pos, int32 width,
+                            int32 *start_pos, int32 width,
                             char *separator, int32 separator_len);
 
 #endif /* NCMPCPP_NC_CYCLIC_BUFFER_H */
